Avoid blocking wait when run outside of story.
This patch uses the StoryMarker to workaround FW-97 and not await module
Initialize() when run outside of the session.

In order to achieve that, we make the Ledger-backed history object queue
up requests until we know if we're running as a module or not. Then we
call History::Initialize() and resolve the pending callbacks.

The practical manifestation of this change is that when running `launch
moterm`, one does not need to wait for 5s before it becomes responsive.

Change-Id: I5ebf8ada387bba6bee63bf1fd988e12d4f7dd01f